# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ",
        "description": "Listado de los productos extras",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/Termo": {
            "get": {
                "tags": [
                    "Termos"
                ],
                "summary": "Consultar todos los termos",
                "description": "Retorna todos los termos disponibles",
                "operationId": "64101dc658f2a0e8b31ac4659d3d0d2f",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Termo térmico"
                                                    },
                                                    "capacidad": {
                                                        "type": "string",
                                                        "example": "500 ml"
                                                    },
                                                    "material": {
                                                        "type": "string",
                                                        "example": "acero inoxidable"
                                                    },
                                                    "precio": {
                                                        "type": "number",
                                                        "example": "25"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Termos"
                ],
                "summary": "Crear un nuevo Termo",
                "description": "Crea un nuevo Termo en el sistema.",
                "operationId": "5d76f7c1402365ed844c9866815bef3b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "Material",
                                    "Capacidad",
                                    "Precio",
                                    "Stock"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Termo de acero inoxidable"
                                    },
                                    "Material": {
                                        "type": "string",
                                        "example": "Acero inoxidable"
                                    },
                                    "Capacidad": {
                                        "type": "integer",
                                        "example": 500
                                    },
                                    "Precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 25.5
                                    },
                                    "Stock": {
                                        "type": "integer",
                                        "example": 200
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Termo creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "Nombre": {
                                            "type": "string",
                                            "example": "Termo de acero inoxidable"
                                        },
                                        "Material": {
                                            "type": "string",
                                            "example": "Acero inoxidable"
                                        },
                                        "Capacidad": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "Precio": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 25.5
                                        },
                                        "Stock": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud incorrecta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/Termo/{id}": {
            "put": {
                "tags": [
                    "Termos"
                ],
                "summary": "Actualizar un Termo existente",
                "description": "Actualiza los detalles de un Termo con el ID proporcionado.",
                "operationId": "696f4d8532b76a719903fb43c30c12ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "Material",
                                    "Capacidad",
                                    "Precio",
                                    "Stock"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Termo de plástico"
                                    },
                                    "Material": {
                                        "type": "string",
                                        "example": "Plástico"
                                    },
                                    "Capacidad": {
                                        "type": "integer",
                                        "example": 750
                                    },
                                    "Precio": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15.99
                                    },
                                    "Stock": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Termo actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "Nombre": {
                                            "type": "string",
                                            "example": "Termo de plástico"
                                        },
                                        "Material": {
                                            "type": "string",
                                            "example": "Plástico"
                                        },
                                        "Capacidad": {
                                            "type": "integer",
                                            "example": 750
                                        },
                                        "Precio": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 15.99
                                        },
                                        "Stock": {
                                            "type": "integer",
                                            "example": 100
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontró el Termo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Termo] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Termos"
                ],
                "summary": "Eliminar un Termo",
                "description": "Elimina un Termo del sistema usando el ID proporcionado.",
                "operationId": "4eda24bbe359826f89797aca98581047",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Termo eliminado correctamente"
                    },
                    "404": {
                        "description": "No se encontró el Termo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Termo] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Sanctum bearer token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearer_token": []
        }
    ]
}